requirements.txt
aioredis==1.3.1
amqp==2.5.2
arrow==0.15.5
asgiref==3.2.3
astroid==2.3.1
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
autopep8==1.4.4
backcall==0.1.0
billiard==3.6.1.0
blessed==1.17.0
celery==4.4.0
certifi==2019.9.11
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.3
constantly==15.1.0
coverage==5.0.1
cryptography==2.8
daphne==2.4.1
decorator==4.4.1
dill==0.3.1.1
Django==2.2.6
django-appconf==1.0.3
django-cors-headers==3.1.1
django-filter==2.2.0
django-grappelli==2.13.2
django-picklefield==2.1.1
django-q==1.1.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
ecdsa==0.14.1
factory-boy==2.12.0
Faker==3.0.0
filelock==3.0.12
funcsigs==1.0.2
future==0.18.2
gunicorn==19.9.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.3.0
incremental==17.5.0
ipdb==0.12.2
ipython==7.10.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
jsonschema==3.2.0
kombu==4.6.7
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
more-itertools==8.0.2
mpu==0.20.0
msgpack==0.6.2
numpy==1.18.1
packaging==20.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pipupgrade==1.6.7
pluggy==0.13.1
prompt-toolkit==3.0.0
protobuf==3.11.3
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.5.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.4.2
pylint-django==2.0.12
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
pytest-django==3.8.0
python-dateutil==2.8.1
python-jose==3.0.1
pytz==2019.2
PyYAML==5.3
ray==0.8.1
redis==3.3.11
requests==2.22.0
rsa==4.0
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.3
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.6
vine==1.3.0
wcwidth==0.1.7
whitenoise==4.1.4
wrapt==1.11.2
zipp==0.6.0
zope.interface==4.7.1
